# backend/openapi/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Contrato HTTP do CRM PNP. Tenant é sempre derivado da credencial.",
    "title" : "CRM PNP API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/conversas/eventos" : {
      "get" : {
        "operationId" : "eventos",
        "parameters" : [ {
          "in" : "query",
          "name" : "apos",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "default" : 0
          }
        }, {
          "in" : "query",
          "name" : "limite",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 100
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaginaEventos"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Requisição ou JSON inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Autenticação ausente ou expirada"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Acesso negado"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Recurso não encontrado"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Conflito de estado ou idempotência"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Regra de negócio ou validação semântica"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Limite de requisições excedido"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Falha interna com detalhe seguro"
          }
        },
        "tags" : [ "conversa-controller" ]
      }
    },
...
    "/api/webhooks/telegram/{channelConnectionId}" : {
      "post" : {
        "operationId" : "receber",
        "parameters" : [ {
          "in" : "path",
          "name" : "channelConnectionId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "header",
          "name" : "X-Telegram-Bot-Api-Secret-Token",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "string",
                "format" : "byte"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Requisição ou JSON inválido"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Autenticação ausente ou expirada"
          },
          "403" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Acesso negado"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Recurso não encontrado"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Conflito de estado ou idempotência"
          },
          "422" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Regra de negócio ou validação semântica"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Limite de requisições excedido"
          },
          "500" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErroResponse"
                }
              }
            },
            "description" : "Falha interna com detalhe seguro"
          }
        },
        "tags" : [ "telegram-webhook-controller" ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "ErroCampo" : {
        "properties" : {
          "campo" : {
            "type" : "string"
          },
          "mensagem" : {
            "type" : "string"
          }
        }
      },
      "ErroResponse" : {
        "properties" : {
          "campos" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ErroCampo"
            }
          },
          "codigo" : {
            "type" : "string"
          },
          "correlacaoId" : {
            "type" : "string"
          },
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "type" : "string",
            "format" : "uri"
          },
          "momento" : {
            "type" : "string",
            "format" : "date-time"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "format" : "uri"
          }
        }
      },
...
      "EventoResposta" : {
        "type" : "object",
        "properties" : {
          "conversationId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "messageId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "ocorridoEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "sequence" : {
            "type" : "integer",
            "format" : "int64"
          },
          "tipo" : {
            "type" : "string"
          },
          "versao" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "PaginaEventos" : {
        "type" : "object",
        "properties" : {
          "eventos" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EventoResposta"
            }
          },
          "resetObrigatorio" : {
            "type" : "boolean"
          },
          "temMais" : {
            "type" : "boolean"
          },
          "ultimaSequencia" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
    }
  }
}